make exploitdb a command after installing the package
